+++ AtlasLoot_BattleforAzeroth/data.lua
@@ -69,18 +69,8 @@ local HEROIC_DUNGEON_DIFF = data:AddDifficulty(AL["Heroic"], "HeroicDungeonWithP
 		addDifficultyBonus = true,
 	},
 }, 2)
-local MYTHICD_DIFF = data:AddDifficulty(AL["Mythic"], "h", nil, 23)
 local MYTHICD_DUNGEON_DIFF = data:AddDifficulty(AL["Mythic"], "MythicDungeonWithPreset", {
 	Item = {
-		item2bonus = "BfAMaxItemLvl",
-		autoCompleteItem2 = true,
-		addDifficultyBonus = true,
-	},
-}, 23)
-local MYTHICD2_DIFF = data:AddDifficulty(AL["Mythic"], "h", nil, 23)
-local MYTHICD2_DUNGEON_DIFF = data:AddDifficulty(AL["Mythic"], "MythicDungeon2WithPreset", {
-	Item = {
-		item1bonus = "LegionMDungeon2",
 		item2bonus = "BfAMaxItemLvl",
 		autoCompleteItem2 = true,
 		addDifficultyBonus = true,